Move transports from CoreRegistrationData to RegistrationData
